VM -> StateManager: moved generateGenesis() to BaseStateManager
